LGTM configuration file (#3859)
Introduce LGTM config file to optimize extraction times. The vanilla setup for LGTM.com analyzes the folder docs which includes the files located in docs/api, where there is ~177KB of minified JS in each html file. If we can avoid this folder during extraction, then there is a huge speedup in analysis time. It does not appear like the code in this folder is of interest for scanning.