Add BigInteger tests (#4520)
* fix:(amazon-cognito-identity-js): include radix parameter

Without this fix, a negative bigint will always end up throwing `new Error('Only radix 2, 4, 8, 16, 32 are supported')` because b is undefined.  This fix is present in the original code from which this file was derived: http://www-cs-students.stanford.edu/~tjw/jsbn/jsbn.js

* Include radix when calling bnToString() for negatives

The bnToString() function throws an exception when called without a radix value or given a radix outside of [2, 4, 8, 16, 32]. 

When bnToString() is called on a negative big int value, it's negated value is used to call bnToString() which then gets concatenated to a '-'. However, in the bnToString() call for the negated value, no radix value is passed in so it always throws an exception. 

This change will use the radix value argument from the initial bnToString() call on the negative big integer value when calling bnToString() for it's negated value so that it works as intended.

* Add BigInteger.test.js

* Update dependencies so that jest runs

* Correct negative test name

* Update .babelrc to mirror previous one